+++ src/components/progress-report/media/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './media.dto';
+export * from './media-list.dto';
+export * from './upload.dto';
