+++ src/components/progress-report/media/dto/media-list.dto.ts
@@ -1,0 +1,30 @@
+import { InputType, ObjectType } from '@nestjs/graphql';
+import {
+  PaginatedList,
+  SortablePaginationInput,
+  Variant,
+  VariantInputField,
+} from '~/common';
+import { ProgressReport } from '../../dto';
+import { MediaVariant, ProgressReportMedia } from './media.dto';
+
+@InputType()
+export class ProgressReportMediaListInput extends SortablePaginationInput<
+  'createdAt' | 'variant'
+>({
+  defaultSort: 'createdAt',
+}) {
+  @VariantInputField(ProgressReportMedia, {
+    nullable: true,
+    many: true,
+    description: 'Filter to these specific variants',
+  })
+  variants?: ReadonlyArray<Variant<MediaVariant>>;
+}
+
+@ObjectType()
+export class ProgressReportMediaList extends PaginatedList(
+  ProgressReportMedia,
+) {
+  readonly report: ProgressReport;
+}
